backend/requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
boto3==1.21.40
botocore==1.24.40
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==36.0.2
cycler==0.11.0
daphne==3.0.2
Deprecated==1.2.13
Django==3.1.14
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
fonttools==4.33.3
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==1.0.0
kiwisolver==1.4.2
kombu==5.2.4
matplotlib==3.5.2
msgpack==1.0.3
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.0
prompt-toolkit==3.0.29
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2021.3
redis==4.2.2
requests==2.27.1
s3transfer==0.5.2
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
Twisted==22.2.0
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zope.interface==5.4.0
